# perf: improve the String concatenation performance