--- a/crates/bench/Cargo.toml
+++ b/crates/bench/Cargo.toml
@@ -14,7 +14,7 @@ doctest = false
 test    = false
 
 [dependencies]
-criterion2         = { version = "2.0.0", default-features = false }
+criterion2         = { workspace = true }
 oxc                = { workspace = true }
 rolldown           = { workspace = true }
 rolldown_sourcemap = { workspace = true }