Considering
```rust
  let base64 = "base64".repeat(1000);

  let mime_ext = MimeExt { mime: "image/png".parse().unwrap(), is_utf8_encoded: true };

  // Mix of String and str
  group.bench_function("mixed_types_concat", |b| {
    let mime_ext_string = mime_ext.to_string();
    b.iter(|| black_box(concat_string!("data:", mime_ext_string, ";base64,", base64)));
  });
  group.bench_function("mixed_types_format", |b| {
    b.iter(|| black_box(format!("data:{mime_ext};base64,{base64}")));
  });
```

`concat_string` is 2x faster than `format!`

```
concat_string/mixed_types_concat
                        time:   [98.035 ns 101.56 ns 105.10 ns]
                        change: [+8.7720% +13.706% +18.797%] (p = 0.00 < 0.05)
                        Performance has regressed.
concat_string/mixed_types_format
                        time:   [217.54 ns 221.50 ns 225.72 ns]
                        change: [+2.2251% +4.5076% +6.7315%] (p = 0.00 < 0.05)
                        Performance has regressed.
```