@hyf0 2024-11-16
Do you think you could merge this PR to your branch? https://github.com/rolldown/rolldown/pull/2764

It's about suggesting people  should use  `concat-string` instead of `format!` if possible.