@hyf0 2024-11-16
You could run `just test update` to update all snapshots both in rust and nodejs.